UPDATE make ItemFilterPrice a SFC
