--- resources/views/ItemList/Components/Filter/ItemFilterPrice.twig
@@ -1,32 +1,0 @@
-<script type="x/template" id="vue-item-filter-price">
-	<div class="item-filter-price">
-		<div class="input-group">
-			<div class="input-group-prepend">
-				<span class="input-group-text">${ currency }</span>
-			</div>
-
-			<input type="number"
-				   class="form-control"
-				   @focus="selectAll($event)"
-				   placeholder="Min"
-				   v-model="priceMin">
-		</div>
-
-		<div class="input-group">
-			<div class="input-group-prepend">
-				<span class="input-group-text">${ currency }</span>
-			</div>
-
-			<input type="number"
-				   class="form-control"
-				   @focus="selectAll($event)"
-				   placeholder="Max"
-				   v-model="priceMax">
-		</div>
-
-		<button type="button" class="btn btn-primary btn-appearance" :class="{'disabled': isDisabled}" v-tooltip
-				data-toggle="tooltip" data-placement="top" @click="triggerFilter()" title="{{ trans("Ceres::Template.itemApply") }}">
-			<i v-waiting-animation="isLoading" class="fa fa-check" aria-hidden="true"></i>
-		</button>
-	</div>
-</script>
